moved map button control file
