+++ front_end/DrawingABlank/src/components/mapView/MapButtons.js
@@ -1,9 +1,9 @@
-import React, {useState} from 'react';
+import React, {Fragment, useState} from 'react';
 import {View, Button} from 'react-native';
-import AbsoluteComponent from '../hocs/AbsoluteComponent';
+import AbsoluteComponent from '../../hocs/AbsoluteComponent';
 import {Icon} from 'react-native-elements';
 
-import {styles} from './style.js';
+import {styles} from '../style.js';
 import {TouchableOpacity} from 'react-native-gesture-handler';
 
 import ModalSelector from 'react-native-modal-selector';
@@ -12,6 +12,8 @@ export default function MapControls({
   toggleGhostMode,
   toggleWorkout,
   toggleShowEventsList,
+  toggleSidebar,
+  refresh_map,
   workout_active,
   ghost_inactive,
 }) {
@@ -26,68 +28,86 @@ export default function MapControls({
   var option_selected = 'walk';
 
   return (
-    <AbsoluteComponent style={{bottom: 70, elevation: 0}}>
-      <View style={styles.buttonContainer}>
-        <TouchableOpacity>
-          {ghost_inactive ? (
+    <Fragment>
+      <AbsoluteComponent style={{left: -20, top: 60}}>
+        <View style={styles.small_buttonContainer}>
+          <TouchableOpacity>
+            <Icon name={'menu'} type={'feather'} onPress={toggleSidebar}></Icon>
+          </TouchableOpacity>
+        </View>
+      </AbsoluteComponent>
+      <AbsoluteComponent style={{top: 60, right: -30}}>
+        <View style={styles.small_buttonContainer}>
+          <TouchableOpacity>
             <Icon
-              name={'eye-off'}
+              name={'refresh-cw'}
+              type={'feather'}
+              onPress={refresh_map}></Icon>
+          </TouchableOpacity>
+        </View>
+      </AbsoluteComponent>
+      <AbsoluteComponent style={{left: 0, right: 0, bottom: 70, elevation: 0}}>
+        <View style={styles.buttonContainer}>
+          <TouchableOpacity>
+            {ghost_inactive ? (
+              <Icon
+                name={'eye-off'}
+                type={'feather'}
+                // iconStyle={styles.icon}
+                // containerStyle={styles.menu}
+                size={30}
+                onPress={toggleGhostMode}
+              />
+            ) : (
+              <Icon
+                name={'eye'}
+                type={'feather'}
+                size={30}
+                onPress={toggleGhostMode}
+              />
+            )}
+          </TouchableOpacity>
+          <TouchableOpacity>
+            {workout_active ? (
+              <Icon
+                name={'pause'}
+                type={'feather'}
+                iconStyle={styles.paused}
+                // containerStyle={styles.menu}
+                size={30}
+                onPress={toggleWorkout} // stop workout (change to stats, -1)
+              />
+            ) : (
+              <ModalSelector
+                data={workout_choices}
+                ref={selector => (this.selector = selector)}
+                customSelector={
+                  <Icon
+                    name={'play'}
+                    type={'feather'}
+                    size={30}
+                    onPress={() => this.selector.open()}
+                  />
+                }
+                onChange={option => {
+                  this.option_selected = option.customKey;
+                  toggleWorkout(this.option_selected);
+                }}
+              /> // start workout(-1, type)
+            )}
+          </TouchableOpacity>
+          <TouchableOpacity>
+            <Icon
+              name={'award'}
               type={'feather'}
               // iconStyle={styles.icon}
               // containerStyle={styles.menu}
               size={30}
-              onPress={toggleGhostMode}
+              onPress={toggleShowEventsList}
             />
-          ) : (
-            <Icon
-              name={'eye'}
-              type={'feather'}
-              size={30}
-              onPress={toggleGhostMode}
-            />
-          )}
-        </TouchableOpacity>
-        <TouchableOpacity>
-          {workout_active ? (
-            <Icon
-              name={'pause'}
-              type={'feather'}
-              iconStyle={styles.paused}
-              // containerStyle={styles.menu}
-              size={30}
-              onPress={toggleWorkout} // stop workout (change to stats, -1)
-            />
-          ) : (
-            <ModalSelector
-              data={workout_choices}
-              ref={selector => (this.selector = selector)}
-              customSelector={
-                <Icon
-                  name={'play'}
-                  type={'feather'}
-                  size={30}
-                  onPress={() => this.selector.open()}
-                />
-              }
-              onChange={option => {
-                this.option_selected = option.customKey;
-                toggleWorkout(-1, this.option_selected);
-              }}
-            /> // start workout(-1, type)
-          )}
-        </TouchableOpacity>
-        <TouchableOpacity>
-          <Icon
-            name={'award'}
-            type={'feather'}
-            // iconStyle={styles.icon}
-            // containerStyle={styles.menu}
-            size={30}
-            onPress={toggleShowEventsList}
-          />
-        </TouchableOpacity>
+          </TouchableOpacity>
 
-        {/* <Button
+          {/* <Button
           title={'Ghost Mode'}
           style={styles.button}
           onPress={toggleGhostMode}></Button>
@@ -99,14 +119,15 @@ export default function MapControls({
           title={'See Events'}
           style={styles.button}
           onPress={toggleShowEventsList}></Button> */}
-        {/* Want a re-centre button, show path as data points button  */}
-      </View>
-      {/* <View style={{top: 20, alignItems: 'center'}}>
+          {/* Want a re-centre button, show path as data points button  */}
+        </View>
+        {/* <View style={{top: 20, alignItems: 'center'}}>
         <Button
           title={'Refresh'}
           style={styles.button}
           onPress={drawGridsFunction}></Button>
       </View> */}
-    </AbsoluteComponent>
+      </AbsoluteComponent>
+    </Fragment>
   );
 }
